// app/(authenticated)/master-record/types.ts
export type CoachMaster = {
    fcoachno: string | null;
    ocoachno: string | null;
    coachno: string;
    code: string | null;
    mfgdt: Date | null;
    yrblt: string | null;
    make: string | null;
    basedepot: string | null;
    re_placed: string | null;
    dt_placed: Date | null;
    comdt: Date | null;
    remaster: string | null;
    books: string | null;
    fund: string | null;
    cost: number | null;
    type: string | null;
    rakesing: string | null;
    tweight: number | null;
    gen: string | null;
    speed: string | null;
    acnac: string | null;
    lhbnlhb: string | null;
    age: number | null;
    overagedt: Date | null;
    aboem: string | null;
    bktype: string | null;
    cbc: string | null;
    cbcby: string | null;
    cbcoem: string | null;
    coupling: string | null;
    corrmhrs: number | null;
    period: number | null;
    lshopoutdt: Date | null;
    lpohby: string | null;
    retpohdt: Date | null;
    age_rt_dt: number | null;
    dt_ioh: Date | null;
    ret_dt_ioh: Date | null;
    mlrdt: Date | null;
    mlrby: string | null;
    date_rf: Date | null;
    date_conv: Date | null;
    seat: number | null;
    berth: number | null;
    lschedule: string | null;
    duesch: string | null;
    ns_ss2_3: string | null;
    status: string | null;
    pohby: string | null;
    yardindt: Date | null;
    shopindt: Date | null;
    allot: string | null;
    rfid_det: string | null;
    shop: string | null;
    rmpu: string | null;
    liftdt: Date | null;
    lowerdt: Date | null;
    reliftdt: Date | null;
    airbkdt: Date | null;
    ptindt: Date | null;
    rfid_dv: string | null;
    rfid_nondv: string | null;
    ptoutdt: Date | null;
    cmm_mdepot: string | null;
    tankfitold: number | null;
    tankfitnew: number | null;
    tankmfg: string | null;
    corrstatus: string | null;
    batteryld: string | null;
    fanload: string | null;
    lightload: string | null;
    utkrisht: string | null;
    lavupgrade: string | null;
    ventury: number | null;
    upholstry: string | null;
    dyno: string | null;
    cmm_depot: string | null;
    h_corr: string | null;
    h_corr_dt: Date | null;
    inissue: Date | null;
    remark: string | null;
    workdays: number | null;
    caldays: number | null;
    dcwifit: Date | null;
    elecracfit: Date | null;
    ntxrfit: Date | null;
    issuedt: string | null;
    innhs1: string | null;
    innhs1dt: Date | null;
    innhs2: string | null;
    innhs2dt: Date | null;
    cond_pos: number | null;
    date_cond: string | null;
    insp_dt: Date | null;
    ioh_by: string | null;
    tnk_c_by: string | null;
    other_3: string | null;
    fiba: string | null;
    asv: string | null;
    wsp: string | null;
    l_3yr_done: Date | null;
    l_6yr_done: Date | null;
    l_9yr_done: Date | null;
    l_3yr_due: Date | null;
    l_6yr_due: Date | null;
    c_r: string | null;
    abk: string | null;
    rl: string | null;
    pt: string | null;
    lif: string | null;
    suspension: string | null;
    elra: string | null;
    l_9yr_due: Date | null;
    uphlddt: Date | null;
    acme: Date | null;
    pfs: string | null;
    uphundt: Date | null;
    btlddt: Date | null;
    btundt: Date | null;
    bt: string | null;
    uph: string | null;
    rly: string | null;
    ctype: string | null;
    roamsdt: Date | null;
    bvt: string | null;
};